Move JEI integration
